Update to JDK8
Now that Sponge is requiring jdk8, we can too! This lets us take
advantage of all sorts of fun functional stuff, allowing code to be
more concise and me to be happier.

+++ permissionsex-core/src/test/java/ninja/leaping/permissionsex/util/command/CommandFlagsTest.java
@@ -38,7 +38,7 @@ public class CommandFlagsTest {
                     @Override
                     public <TextType> void execute(Commander<TextType> src, CommandContext args) throws CommandException {
                         assertEquals(true, args.getOne("a"));
-                        assertEquals(42, args.getOne("quot"));
+                        assertEquals((Integer) 42, args.getOne("quot"));
                         assertEquals("something", args.getOne("key"));
                     }
                 })
